requirements.txt
fastapi>=0.68.0,<0.69.0
uvicorn>=0.15.0,<0.16.0
sqlalchemy>=1.4.0,<1.5.0
pydantic>=1.8.0,<1.9.0
email-validator>=1.1.3,<1.2.0
psycopg2-binary>=2.9.1,<2.10.0
alembic>=1.7.4,<1.8.0
aioredis>=2.0.0,<2.1.0
redis>=4.0.0,<5.0.0
python-dotenv>=0.19.0,<0.20.0
pytest>=6.2.5,<6.3.0
httpx>=0.19.0,<0.20.0
google-api-python-client>=2.0.0,<3.0.0
google-auth-httplib2>=0.1.0,<0.2.0
google-auth-oauthlib>=0.4.0,<0.5.0
twilio>=7.0.0,<7.1.0
celery>=5.2.0,<5.3.0
flower>=1.0.0,<1.1.0
